Add student on global level
A student is added on a global level with its uid, name, and array of courses. This will be used to verify student UID on login, for creating StudentCourseSelector (seeing different courses), and allows professors to safely delete students and courses which effects which courses the student can see.

If the student has no more courses, the student gets delete on a global level, making its UID invalid since it's not in any courses